ADD Max to Counter for security

+++ programs/fundraising-protocol/src/state/project.rs
@@ -11,7 +11,7 @@ pub struct Project {
     pub description: String,
     pub funding_goal: u64, // Funding goal in lamports
     pub deadline: i64,     // Deadline timestamp
-    pub project_id: u64,   // Project id (from counter)
+    pub project_id: u8,   // Project id (from counter)
     pub is_success: bool,
     pub is_finalized: bool,
     pub bump: u8, // PDA Bump seed
